refactor(explore): preserve control-plane boundary
